add msvc support. using assimp library import model

+++ include/core/clipper.h
@@ -20,7 +20,7 @@ public:
 
 protected:
     Triangle *triangle;
-    int primitiveCount;
+    size_t primitiveCount;
     bool canClip;
 };
